Revert "LPS-137761 Make window available methods created from Editor component"
This reverts commit 36fca08c697009202c8dcc31fa63d0950f1f1e37.
